added: disabling edit if shared on read only

+++ src/app/pages/list-details/list-details.page.ts
@@ -1,3 +1,4 @@
+import { AuthService } from './../../services/auth.service';
 import { CreateTodoComponent } from './../../modals/create-todo/create-todo.component';
 import { ToastController, ModalController } from '@ionic/angular';
 import { ListService } from './../../services/list.service';
@@ -5,8 +6,8 @@ import { Component, OnInit } from '@angular/core';
 import { List } from 'src/app/models/list';
 import { ActivatedRoute, Route, Router } from '@angular/router';
 import { Todo } from 'src/app/models/todo';
-import {TodoService} from '../../services/todo.service';
-import {Observable} from 'rxjs';
+import { TodoService } from '../../services/todo.service';
+import { merge, Observable, combineLatest } from 'rxjs';
 import { map, switchMap, tap } from 'rxjs/operators';
 
 @Component({
@@ -17,58 +18,74 @@ import { map, switchMap, tap } from 'rxjs/operators';
 export class ListDetailsPage implements OnInit {
   list: Observable<List>;
   listId: string;
+  owner: boolean = false;
+  canWrite: boolean = false;
 
-
-  constructor(private listService: ListService, private route: ActivatedRoute, private router: Router, private toast: ToastController, private modalController: ModalController, private todoService: TodoService) { }
+  constructor(private listService: ListService, private route: ActivatedRoute, private router: Router, private toast: ToastController, private modalController: ModalController, private todoService: TodoService, private auth: AuthService) { }
 
   ngOnInit() {
     const listId = this.route.snapshot.paramMap.get('listId');
     this.list = this.listService.getOneObs(listId).pipe(
       switchMap((l: List) => this.todoService.getListTodos(l).pipe(
-          map((todos) => l.todos = todos),
-          map((_) => l.todos.sort((a: Todo, b: Todo) => a.createdAt - b.createdAt)),
-          map((_) => l),
-        ))
+        map((todos) => l.todos = todos),
+        map((_) => l.todos.sort((a: Todo, b: Todo) => a.createdAt - b.createdAt)),
+        map((_) => l),
+      ))
     );
 
-    this.list.subscribe((list) => {
+    let currentUid$ = this.auth.getConnectedUser().pipe(map((user) => user.uid));
+    combineLatest([currentUid$, this.list]).subscribe(([uid, list]) => {
+      this.owner = list.owner == uid;
+      this.canWrite = this.owner;
+      if (!this.owner) {
+        this.canWrite = list.canWrite.includes(uid);
+      }
+
       this.listId = list.id;
     });
   }
 
-  back(): void{
+  back(): void {
     this.router.navigateByUrl('/home');
   }
 
-  async newTodoModal(todo?: Todo){
-    const modal = await this.modalController.create({
-      component: CreateTodoComponent,
-      cssClass: 'create-todo',
-      componentProps: {
-        listId: this.listId,
-        todo: todo,
-      }
-    });
+  async newTodoModal(todo?: Todo) {
+    if (this.canWrite) {
+      const modal = await this.modalController.create({
+        component: CreateTodoComponent,
+        cssClass: 'create-todo',
+        componentProps: {
+          listId: this.listId,
+          todo: todo,
+        }
+      });
 
-    return await modal.present();
-  }
-
-  updateTodo(todo: Todo) : void {
-    this.newTodoModal(todo);
-  }
-
-  deleteTodo(todo: Todo): void {
-    if (confirm("This Todo task will be deleted, are you sure?")) {
-      this.todoService.delete(todo.id, this.listId);
+      return await modal.present();
     }
   }
 
-  goToTodo(todo: Todo): void{
-      this.router.navigateByUrl('/lists/' + this.listId + '/todos/' + todo.id);
+  updateTodo(todo: Todo): void {
+    if (this.canWrite) {
+      this.newTodoModal(todo);
+    }
+  }
+
+  deleteTodo(todo: Todo): void {
+    if (this.canWrite) {
+      if (confirm("This Todo task will be deleted, are you sure?")) {
+        this.todoService.delete(todo.id, this.listId);
+      }
+    }
+  }
+
+  goToTodo(todo: Todo): void {
+    this.router.navigateByUrl('/lists/' + this.listId + '/todos/' + todo.id);
   }
 
   toggleTodoDone(todo: Todo) {
-    todo.isDone = !todo.isDone;
-    this.todoService.update(todo, this.listId);
+    if (this.canWrite) {
+      todo.isDone = !todo.isDone;
+      this.todoService.update(todo, this.listId);
+    }
   }
 }
